+++ examples/data/IC160A.json
@@ -1,0 +1,218 @@
+{
+    "name": "HADES Inverted Coax",
+    "units": {
+        "length": "mm",
+        "angle": "deg",
+        "potential": "V",
+        "temperature": "K"
+    },
+    "grid": {
+        "coordinates": "cylindrical",
+        "axes": {
+            "r": {
+                "to": 40.0,
+                "boundaries": "inf"
+            },
+            "phi": {
+                "from": 0,
+                "to":0,
+                "boundaries": "periodic"
+            },
+            "z": {
+                "from": -10,
+                "to": 90.0,
+                "boundaries": {
+                    "left": "inf",
+                    "right": "inf"
+                }
+            }
+        }
+    },
+    "medium": "vacuum",
+    "objects": [
+        {
+            "type": "semiconductor",
+            "material": "HPGe",
+            "bulk_type": "p",
+            "temperature": 77.0,
+            "charge_density_model": {
+                "name": "linear",
+                "r": {
+                    "init": 0,
+                    "gradient": 0
+                },
+                "phi": {
+                    "init": 0.0,
+                    "gradient": 0.0
+                },
+                "z": {
+                    "init": -1e7,
+                    "gradient": 1e5
+                }
+            },
+            "geometry": {
+                "type": "difference",
+                "parts": [
+                    {
+                        "name": "Initial Cylinder",
+                        "type": "tube",
+                        "r": {
+                            "from": 0.0,
+                            "to": 35.0
+                        },
+                        "phi": {
+                            "from": 0.0,
+                            "to": 360.0
+                        },
+                        "h": 80.0
+                    },
+                    {
+                        "name": "borehole",
+                        "type": "cone",
+                        "r": {
+                            "bottom": {
+                                "from": 0.0,
+                                "to": 5.0
+                            },
+                            "top": {
+                                "from": 0.0,
+                                "to": 10.0
+                            }
+                        },
+                        "phi": {
+                            "from": 0.0,
+                            "to": 360.0
+                        },
+                        "h": 55.0,
+                        "translate": {
+                            "z": 25.0
+                        }
+                    },
+                    {
+                        "name": "ditch",
+                        "type": "tube",
+                        "r": {
+                            "from": 10.0,
+                            "to": 15.0
+                        },
+                        "phi": {
+                            "from": 0.0,
+                            "to": 360.0
+                        },
+                        "h": 5.0
+                    }
+                ]
+            }
+        },
+
+        {
+            "type": "contact",
+            "material": "HPGe",
+            "channel": 1,
+            "potential": 0.0,
+            "geometry": {
+                "type": "tube",
+                "r": {
+                    "from": 0.0,
+                    "to": 1.5
+                },
+                "phi": {
+                    "from": 0.0,
+                    "to": 360.0
+                },
+                "h": 0.5
+            }
+        },
+        {
+            "type": "contact",
+            "material": "HPGe",
+            "channel": 2,
+            "potential": 4000.0,
+            "geometry": {
+                "type": "union",
+                "parts":  [
+                    {
+                        "name": "bottom lid",
+                        "type": "tube",
+                        "r": {
+                            "from": 15.0,
+                            "to": 35.0
+                        },
+                        "phi": {
+                            "from": 0.0,
+                            "to": 360.0
+                        },
+                        "h": 0.0
+                    },
+                    {
+                        "name": "side linning",
+                        "type": "tube",
+                        "r": {
+                            "from": 35.0,
+                            "to": 35.0
+                        },
+                        "phi": {
+                            "from": 0.0,
+                            "to": 360.0
+                        },
+                        "h": 80.0
+                    },
+                    {
+                        "name": "top lid",
+                        "type": "tube",
+                        "r": {
+                            "from": 10.0,
+                            "to": 35
+                        },
+                        "phi": {
+                            "from": 0.0,
+                            "to": 360.0
+                        },
+                        "h": 0.0,
+                        "translate": {
+                            "z": 80.0
+                        }
+                    },
+                    {
+                        "name": "borehole linning",
+                        "type": "cone",
+                        "r": {
+                            "bottom": {
+                                "from": 5.0,
+                                "to": 5.0
+                            },
+                            "top": {
+                                "from": 10.0,
+                                "to": 10.0
+                            }
+                        },
+                        "phi": {
+                            "from": 0.0,
+                            "to": 360.0
+                        },
+                        "h": 55.0,
+                        "translate": {
+                            "z": 25.0
+                        }
+                    },
+                    {
+                        "name": "borehole bottom lid",
+                        "type": "tube",
+                        "r": {
+                            "from": 0.0,
+                            "to": 5.0
+                        },
+                        "phi": {
+                            "from": 0.0,
+                            "to": 360.0
+                        },
+                        "h": 0.0,
+                        "translate": {
+                            "z": 25.0
+                        }
+                    }
+                ]
+            }
+        }
+    ]
+}
