+++ src/mcstp_to_mcpss.jl
@@ -12,6 +12,19 @@ Random.seed!(123) # only for testing
 # has to be greater than DAQ waveform length (daq_nsamples in mcpss_to_t1pss.jl)
 # total_waveform_length = 8000;
 
+function mcstp_to_mcpss(det_path::AbstractString, det_name::AbstractString, mc_events::Table, sim_config_file::AbstractString)
+    sim_config = LegendGeSim.load_config(sim_config_file)
+    mcstp_to_mcpss(det_path, det_name, mc_events, sim_config)
+
+end
+
+
+function mcstp_to_mcpss(det_path::AbstractString, det_name::AbstractString, mc_events::Table, sim_config::PropDict)
+    noise_model = NoiseModel(sim_config)
+    mcstp_to_mcpss(det_path, det_name, mc_events, noise_model)
+end
+
+
 ##
 """
     cstp_to_mcpss(det_path, det_name, mc_events)
@@ -26,23 +39,20 @@ mc_events: table in the mcstp format (output of g4_to_mcstp)
 
 Output: Table, Table
 """
-function mcstp_to_mcpss(det_path::AbstractString, det_name::AbstractString, mc_events::Table, sim_config::PropDict)
+function mcstp_to_mcpss(det_path::AbstractString, det_name::AbstractString, mc_events::Table, noise_model::NoiseModel)
 
     simulation = detector_simulation(det_path, det_name)
 
-    mcstp = MCstp(simulation, mc_events)
-
     # add fano noise, don't add if data noise is applied later
-    if(sim_config.noise_data == 0)
-        add_fnoise(mcstp)
-    end
+    mc_events = add_noise(mc_events, simulation, noise_model)
 
     # simulate waveforms
-    mcpss_table, mcpss_mctruth = simulate_wf(mcstp, SSDSimulator())
+    mcpss_table, mcpss_mctruth = simulate_wf(mc_events, simulation, SSDSimulator())
 
     mcpss_table, mcpss_mctruth
 
 end
+
 
 """
     detector_simulation(det_path, det_name)
@@ -119,70 +129,29 @@ end
 
 
 """
-    read_mcstp(filename)
-
-Helper function to read MC events from an HDF5 file with mcstp format
-(g4simple events in the format compatible with SSD simulate_waveforms method)
-(to be rewritten as a g4_to_mcstp() function that reads from storage)
-
-filename: name of mcstp file
-Output: Table
-"""
-function read_mcstp(mcfilename::AbstractString)
-    # Let's read in some Monte-Carlo events (produced by Geant4).
-    # We'll either read from Geant4 CSV and cache the result as HDF5,
-    # or read directly from HDF5 if already available:
-
-    mctruth_filename_csv = "data/$(mcfilename)_mcstp.csv"
-    mctruth_filename_hdf5 = "cache/$(mcfilename)_mcstp.h5"
-
-    if isfile(mctruth_filename_hdf5)
-        @info("Reading MC events from HDF5.")
-        mc_events = HDF5.h5open(mctruth_filename_hdf5, "r") do input
-            readdata(input, "mctruth")
-        end
-    else
-        @info("Reading MC events from Geant4-CSV.")
-        mc_events = open(read, mctruth_filename_csv, Geant4CSVInput)
-        mkpath(dirname(mctruth_filename_hdf5))
-        HDF5.h5open(mctruth_filename_hdf5, "w") do output
-            writedata(output, "mctruth", mc_events)
-        end
-    end
-
-    mc_events
-
-end
-
-##
-"""
-    MCstp(simulation, mc_events)
-A struct containing two things needed for waveform simulation:
-    simulation: SSD detector simulation object
-    mc_events: Table of mcstp format
-"""
-mutable struct MCstp
-    simulation::SolidStateDetectors.Simulation
-    mc_events::Table
-end
-
-"""
-    add_fnoise(mcstp)
+    add_fnoise(mc_events, simulation, noise_model)
 
 Add fano noise to MC events
 
-mcstp: MCstp struct
 
 Output: no output, the function modifies the field mc_events
 """
-function add_fnoise(mcstp::MCstp)
+function add_noise(mc_events::Table, simulation::SolidStateDetectors.Simulation, noise_model::NoiseSim)
     @info("...adding fano noise")
-    det_material = mcstp.simulation.detector.semiconductors[1].material
-    mcstp.mc_events = add_fano_noise(mcstp.mc_events, det_material.E_ionisation, det_material.f_fano)
+    det_material = simulation.detector.semiconductors[1].material
+    add_fano_noise(mc_events, det_material.E_ionisation, det_material.f_fano)
 end
 
+
+function add_noise(mc_events::Table, simulation::SolidStateDetectors.Simulation, noise_model::NoiseData)
+    # do nothing since if we're using noise from data we do not simulate fano noise
+    # not to double count
+    mc_events
+end
+
+
 """
-    simulate_wf(mcstp, simulator)
+    simulate_wf(mc_events, simulation, simulator)
 
 Simulate waveforms based on given MC events and detector simulation,
 contained in the MCstp struct. Returns resulting mcpss table
@@ -193,20 +162,20 @@ simulator: Simulator object inheriting from PSSimulator for multiple dispatch
 
 Output: Table, Table
 """
-function simulate_wf(mcstp::MCstp, ::SSDSimulator)
+function simulate_wf(mc_events::Table, simulation::SolidStateDetectors.Simulation, ::SSDSimulator)
     # We need to filter out the few events that,
     # due to numerical effects, lie outside of the detector
     # (the proper solution is to shift them slightly, this feature will be added
     # in the future):
 
-    filtered_events = mcstp.mc_events[findall(pts -> all(p -> T.(ustrip.(uconvert.(u"m", p))) in mcstp.simulation.detector, pts), mcstp.mc_events.pos)];
+    filtered_events = mc_events[findall(pts -> all(p -> T.(ustrip.(uconvert.(u"m", p))) in simulation.detector, pts), mc_events.pos)];
 
     @info("Simulating waveforms")
     contact_charge_signals = SolidStateDetectors.simulate_waveforms(
             # filtered_events[1:2000],
             # filtered_events[1:800],
             filtered_events,
-            mcstp.simulation,
+            simulation,
             max_nsteps = 4000,
             Δt = 1u"ns",
             verbose = false);
