--- src/utils.jl
@@ -1,3 +1,0 @@
-# This file is a part of LegendGeSim.jl, licensed under the MIT License (MIT).
-
-dummy_function(x::Real) = x^2
